Extract the Variables backend out of the Debug module
Fixes #33
