+++ test/Main.hs
@@ -7,7 +7,6 @@
 module Main(main) where
 
 import Debug
-import Debug.Record
 import Debug.Util
 import Data.List
 import Control.Exception.Extra
